Refactor utils library for state machine checks
