+++ qc_otx/checks/state_machine_checker/__init__.py
@@ -2,5 +2,3 @@ from . import state_machine_constants as state_machine_constants
 from . import state_machine_checker as state_machine_checker
 from . import no_procedure_realization as no_procedure_realization
 from . import mandatory_target_state as mandatory_target_state
-from . import models as models
-from . import utils as utils
